Предупреждения компилятора

+++ sources/VS/Editor/src/Dialogs/TrapezeDialog.cpp
@@ -156,8 +156,8 @@ static void SendForm()
 
     int pointsInTrapeze = static_cast<int>(Point::NUM_POINTS) - delay;
 
-    int vertex1 = center + pointsInTrapeze / 2.0F * scVertex1->GetValue() / 100.0F;
-    int vertex2 = center + pointsInTrapeze / 2.0F * scVertex2->GetValue() / 100.0F;
+    int vertex1 = static_cast<int>(center + pointsInTrapeze / 2.0F * scVertex1->GetValue() / 100.0F);
+    int vertex2 = static_cast<int>(center + pointsInTrapeze / 2.0F * scVertex2->GetValue() / 100.0F);
 
     int levelHI = static_cast<int>(Point::AVE_VALUE - (Point::MAX_VALUE + Point::MIN_VALUE) / 2.0F * scUp->GetValue() / 100.0F); //-V2007
     int levelLOW = static_cast<int>(Point::AVE_VALUE - (Point::MAX_VALUE + Point::MIN_VALUE) / 2.0F * scDown->GetValue() / 100.0F); //-V2007
